refactor: extract camera photo processing
